MNT: Move olog callback into callbacks.olog
