+++ bluesky/callbacks/core.py
@@ -541,34 +541,3 @@ class LiveRaster(CallbackBase):
             self.im.set_clim(np.nanmin(self._Idata), np.nanmax(self._Idata))
 
         self.im.set_array(self._Idata)
-
-
-class OlogCallback(CallbackBase):
-    """Example callback to customize the logbook.
-
-    This is not necessarily the best example of how to customize the log book,
-    but it is something that fits the needs of CHX
-
-    Example
-    -------
-    # add this callback to the run engine
-    >>> gs.RE.subscribe('start', OlogCallback())
-    # turn off the default logger
-    >>> gs.RE.logbook = None
-    """
-    def __init__(self):
-        # import a whole mess of stuff when this thing gets created
-        from pyOlog import SimpleOlogClient
-        self.client = SimpleOlogClient()
-
-    def start(self, doc):
-        from IPython import get_ipython
-        commands = list(get_ipython().history_manager.get_range())
-        document_content = ('%s: %s\n\n'
-                            'RunStart Document\n'
-                            '-----------------\n'
-                            '%s' % (doc['scan_id'],
-                                    commands[-1][2],
-                                    pformat(doc)))
-        olog_status = self.client.log(document_content, logbooks='Data Acquisition')
-        print('client.log returned %s' % olog_status)
